HouseBuilder refactored

+++ CommunalCalculatorTest/Models/Rates/CommunalRate.cs
@@ -3,6 +3,6 @@
     internal class CommunalRate
     {
         internal decimal Rate { get; set; }
-        internal double Normative { get; set; }
+        internal decimal Normative { get; set; }
     }
 }
